+++ modules/apis/src/test/java/com/axway/apim/cli/APIExportCLIOptionsTest.java
@@ -28,7 +28,7 @@ public class APIExportCLIOptionsTest {
 		APIExportParams params = (APIExportParams) options.getParams();
 		Assert.assertEquals(params.getUsername(), "apiadmin");
 		Assert.assertEquals(params.getPassword(), "changeme");
-		Assert.assertEquals(params.getHostname(), "api-env");
+		Assert.assertEquals(params.getHostname(), "localhost");
 		
 		Assert.assertEquals(params.getWide(), Wide.wide);
 		Assert.assertTrue(params.isDeleteTarget());
@@ -54,7 +54,7 @@ public class APIExportCLIOptionsTest {
 		APIExportParams params = (APIExportParams) options.getParams();
 		Assert.assertEquals(params.getUsername(), "apiadmin");
 		Assert.assertEquals(params.getPassword(), "changeme");
-		Assert.assertEquals(params.getHostname(), "api-env");
+		Assert.assertEquals(params.getHostname(), "localhost");
 		
 		Assert.assertEquals(params.getWide(), Wide.ultra);
 		// Validate target is current directory if not given
@@ -69,7 +69,7 @@ public class APIExportCLIOptionsTest {
 		// Validate core parameters are included
 		Assert.assertEquals(params.getUsername(), "apiadmin");
 		Assert.assertEquals(params.getPassword(), "changeme");
-		Assert.assertEquals(params.getHostname(), "api-env");
+		Assert.assertEquals(params.getHostname(), "localhost");
 		
 		// Validate wide is is using standard as default
 		Assert.assertEquals(params.getWide(), Wide.standard);
@@ -93,7 +93,7 @@ public class APIExportCLIOptionsTest {
 		// Validate core parameters are included
 		Assert.assertEquals(params.getUsername(), "apiadmin");
 		Assert.assertEquals(params.getPassword(), "changeme");
-		Assert.assertEquals(params.getHostname(), "api-env");
+		Assert.assertEquals(params.getHostname(), "localhost");
 		
 		// Validate an API-Filter parameters are included
 		Assert.assertEquals(params.getApiPath(), "/api/v1/greet");
@@ -110,7 +110,7 @@ public class APIExportCLIOptionsTest {
 		// Validate core parameters are included
 		Assert.assertEquals(params.getUsername(), "apiadmin");
 		Assert.assertEquals(params.getPassword(), "changeme");
-		Assert.assertEquals(params.getHostname(), "api-env");
+		Assert.assertEquals(params.getHostname(), "localhost");
 		
 		// Validate an API-Filter parameters are included
 		Assert.assertEquals(params.getApiPath(), "/api/v1/to/be/upgraded");
@@ -141,7 +141,7 @@ public class APIExportCLIOptionsTest {
 		// Validate core parameters are included
 		Assert.assertEquals(params.getUsername(), "apiadmin");
 		Assert.assertEquals(params.getPassword(), "changeme");
-		Assert.assertEquals(params.getHostname(), "api-env");
+		Assert.assertEquals(params.getHostname(), "localhost");
 		
 		// Validate an API-Filter parameters are included
 		Assert.assertEquals(params.getApiPath(), "/api/v1/some");